[IMP] prices creation method

+++ app/controllers/events/events_controller.ts
@@ -19,9 +19,9 @@ import { createMediaValidator } from '#validators/media'
 import { createPricesValidator } from '#validators/price'
 import { queryValidator } from '#validators/query'
 import User from '#models/user'
-import { updatePricesValidator } from '#validators/update_price'
 import app from '@adonisjs/core/services/app'
 import { cuid } from '@adonisjs/core/helpers'
+import { createPriceValidator } from '#validators/create_price'
 
 export default class EventsController {
   /**
@@ -141,7 +141,7 @@ export default class EventsController {
   /**
    * Handle form submission for the create action
    */
-  async store({ request, session, response, auth }: HttpContext) {
+  async store({ request, session, response, i18n , auth }: HttpContext) {
     const event = await this.createEvent(request, session, response)
     if (event) {
       const user = auth.user
@@ -149,7 +149,7 @@ export default class EventsController {
       await this.attachCategoryTypes(request, session, event)
       await this.attachIndicators(request, event)
       await this.createEventAddress(request, event)
-      await this.createEventPrices(request, event)
+      if (!(await this.createEventPrices(request, session, response, i18n, event))) return response.redirect().back()
       await this.uploadEventMedia(request, event)
 
       return response.redirect().toRoute('events.show', { id: event.id })
@@ -355,25 +355,48 @@ export default class EventsController {
     }
   }
 
-  async createEventPrices(request: HttpContext['request'], event: Event) {
-    // TODO create 5 prices per event
-    // try {
-    const pricePayloads = await request.validateUsing(createPricesValidator)
+  async createEventPrices(
+    request: HttpContext['request'],
+    session: HttpContext['session'],
+    response: HttpContext['response'],
+    i18n: HttpContext['i18n'],
+    event: Event
+  ) {
+    const bodyPrices = request.body().prices
 
-    for (const pricePayload of pricePayloads.prices) {
-      const price = new Price()
+    if (bodyPrices) {
+      // we process one price element at a time
+      bodyPrices.forEach(async (priceData: any) => {
 
-      if (pricePayload.price_description) price.description = pricePayload.price_description
-      if (pricePayload.regular_price) price.regularPrice = pricePayload.regular_price
-      if (pricePayload.discounted_price) price.discountedPrice = pricePayload.discounted_price
-      if (pricePayload.available_qty) price.availableQty = pricePayload.available_qty
-
-      await price.save()
-      await price.related('event').associate(event)
+        try {
+          const payload = await createPriceValidator.validate(priceData)
+  
+          if (payload) {
+            const price = new Price
+            if (payload.price_description) price.description = payload.price_description
+            if (payload.regular_price) price.regularPrice = payload.regular_price
+            if (payload.discounted_price) price.discountedPrice = payload.discounted_price
+            if (payload.available_qty) price.availableQty = payload.available_qty
+  
+            await price.save()
+            await price.related('event').associate(event)
+          }
+        } catch (error) {
+          let errorMsg = i18n.t('messages.errorCreatePrice') + ' '
+          error.messages.forEach((msg: string) => {
+            errorMsg += msg
+          })
+          session.flash('errorCreatePrice', errorMsg)
+          return false
+        }
+      })
+      return true
+    } else {
+      // TODO this error message is not displayed when no prices are entered (payload empty)
+      const errorMsg = i18n.t('messages.errorMissingPrices') + ' '
+      session.flash('errorMissingPrices', errorMsg)
     }
-    // } catch (error) {
-    // console.error('Price Validation Error at createEventPrices():', error)
-    // }
+    return false
   }
 
   async createEventAddress(request: HttpContext['request'], event: Event) {
@@ -473,7 +496,7 @@ export default class EventsController {
           media.path = `/uploads/${uniqueFileName}`
           media.altName = file.clientName
           media.eventId = event.id
-  
+
           await media.save()
         }
       } catch (error) {
@@ -486,21 +509,18 @@ export default class EventsController {
   }
 
   async deleteEventMedia({ params, response, session, i18n }: HttpContext) {
-    console.log('\n\n\n\n\n\n', params)
     const media = await Media.find(params['id'])
-    console.log('media: ', media);
-    try {
-      await media.delete()
-    } catch (error) {
-      console.log(error)
-      // const errorMsg = i18n.t('messages.errorDestroyEvent')
-      // session.flash('error', errorMsg)
+    if (media) {
+      try {
+        await media.delete()
+      } catch (error) {
+        console.log(error)
+        const errorMsg = i18n.t('messages.errorDestroyEvent')
+        session.flash('error', errorMsg)
+      }
+      response.redirect().back()
     }
-    response.redirect().back()
-    // const successMsg = i18n.t('messages.successDestroyEvent')
-    // session.flash('success', successMsg)
   }
-
 
   async getTodayEvents() {
     // get 5 events that occur today
@@ -679,8 +699,8 @@ export default class EventsController {
     i18n: HttpContext['i18n'],
     event: Event
   ) {
-    try {
-      const addressPayload = await request.validateUsing(createAddressValidator)
+    const addressPayload = await request.validateUsing(createAddressValidator)
+    if (addressPayload) {
       event.location.name = addressPayload.name ?? ''
       event.location.street = addressPayload.street.replace('&#x27;', "'")
       event.location.number = addressPayload.number
@@ -701,11 +721,8 @@ export default class EventsController {
 
       await event.location.save()
       return true
-    } catch (error) {
-      const errorMsg = i18n.t('messages.errorMissingAddress') + ' ' + error
-      session.flash('errorMissingAddress', errorMsg)
-      return false
     }
+    return false
   }
 
   async updateEventCategoryTypes(
@@ -774,38 +791,44 @@ export default class EventsController {
     event: Event
   ) {
     // TODO << Investigate >>
+    // [
+    //   {
+    //     price_description: [ 'adultes', 'enfants' ],
+    //     available_qty: [ '5', '3' ],
+    //     regular_price: [ '10', '10' ],
+    //     discounted_price: [ '5', '3' ]
+    //   }
+    // ]
     // prices update can work successfully multiple times in a row
     // then after a while the validation stops working and price update stops working
-    try {
-      const pricePayloads = await request.validateUsing(updatePricesValidator)
-      
-      // if data validation if ok, we delete all prices associated with the event
+    console.log('\n\n\n\n')
+    const data = request.all().prices
+    console.log('data: ', data);
+
+    // const pricePayloads = await request.validateUsing(createPricesValidator)
+
+    if (false) {
+      // we delete all prices associated with the event
       const prices = event.prices
       prices.forEach((price: Price) => {
         price.delete()
       })
 
       // then we create new prices objects and associate them with the event
-      for (const pricePayload of pricePayloads.prices) {
-        const price = new Price()
+      // for (const pricePayload of pricePayloads.prices) {
+      //   const price = new Price()
 
-        if (pricePayload.price_description) price.description = pricePayload.price_description
-        if (pricePayload.regular_price) price.regularPrice = pricePayload.regular_price
-        if (pricePayload.discounted_price) price.discountedPrice = pricePayload.discounted_price
-        if (pricePayload.available_qty) price.availableQty = pricePayload.available_qty
+      //   if (pricePayload.price_description) price.description = pricePayload.price_description
+      //   if (pricePayload.regular_price) price.regularPrice = pricePayload.regular_price
+      //   if (pricePayload.discounted_price) price.discountedPrice = pricePayload.discounted_price
+      //   if (pricePayload.available_qty) price.availableQty = pricePayload.available_qty
 
-        await price.save()
-        await price.related('event').associate(event)
-      }
-
+      //   await price.save()
+      //   await price.related('event').associate(event)
+      // }
       return true
-
-    } catch (error) {
-      console.error('Validation Error:', error.messages) 
-      const errorMsg = i18n.t('messages.errorEditEventPrices') + ' ' + error
-      session.flash('errorEditEventPrices', errorMsg)
-      return false
     }
+    return true
   }
 
   /**
@@ -877,8 +900,8 @@ export default class EventsController {
 
       if (!(await this.updateEventAddress(request, session, i18n, event))) return response.redirect().back()
       if (!(await this.updateEventCategoryTypes(request, session, i18n, event))) return response.redirect().back()
-      await this.updateEventIndicators(request, event)
-      // if (!(await this.updateEventPrices(request, session, i18n, event))) return response.redirect().back()
+      await this.updateEventIndicators(request, event)  // Indicators are optional
+      if (!(await this.updateEventPrices(request, session, i18n, event))) return response.redirect().back()
       if (!(await this.updateEventMedia(request, session, i18n, event))) return response.redirect().back()
 
       await event.save()
